[Sidebar Refactor] Move component inside header
