+++ lib/kits/core-ui/components/paged-publication/page-list.styl
@@ -16,6 +16,7 @@
 
     > img
         width 100%
+        border-radius 8px
 
 @media (max-width: 450px)
     .sgn-pages-list-item-container
